eager load with const_get [closes #58]
This ensures autoloading and eager loading are consistent, specially
with regard to detecting wrong constant names in some edge cases.

For example, in the test included in this patch, Zeitwerk did not know
it was managing cli/x.rb, because it descends lazily. Therefore, that
file defined the CLI module just fine, and the require call did not
trigger the callback that validates the expected constant.