Update README
Add a heading "Setting a custom namespace explicitly" under
"File structure" because it's a distinct feature so it deserves its own
section. This way it will be easier to spot it. I was looking for this
feature specifically in the README and missed it, because I clicked on
"Explicit namespaces" header and it was not there, so I assumed it's not
supported.