Logs autoloads discarded due to implicit -> explicit
If a directory for an explicit namespace is found before its file, you get a
trace

    autoload set for N, to be autovivified from <directory>

Then, when the file is found, you also see

    autoload set for N, to be loaded from <file>

and there's nothing connecting the two of them, except for the constant path,
which is the same ("N", in the example).

This new message makes the namespace promotion more clear.